requirements.txt
flask==2.3.3
openai==1.3.8
pinecone-client==3.0.0
sentence-transformers==2.2.2
torch==2.1.0
transformers==4.35.2
numpy==1.24.3
scikit-learn==1.3.2
gunicorn==21.2.0